requirements.txt
Flask==3.1.0
flask-cors==5.0.0
tensorflow-cpu==2.17.0
keras==3.8.0
Pillow==10.2.0
numpy==1.26.4
gunicorn==21.2.0
